Sharding-aware reconciliation
This does two things:
1. Allows reconciliation to be skipped when adding peers.
2. Makes the reconciliation sharding-aware by only reconciling covalues
   that are relevant to the peer being reconciled.

(1) is necessary because relevance is determined based on the current
peer set, so if we always reconciled when adding peers, the first peer
would always be deemed relevant for every single covalue. Similar but
different problem for subsequent peers: there would be a ton of covalue
rebalancing as peers were added.

So, the following is the expected pattern when setting up a node with
sharded server peers:
1. Determine the full set of peers upfront
2. Add them all without reconciliation
3. Reconcile them all